+++ src/udp_lat.c
@@ -42,7 +42,7 @@
 #endif
 
 int main(int argc, char *argv[]) {
-  int size;
+  ssize_t size;
   char *buf;
   int64_t count, i, delta;
 #ifdef HAS_CLOCK_GETTIME_MONOTONIC
@@ -52,7 +52,7 @@ int main(int argc, char *argv[]) {
 #endif
 
   ssize_t len;
-  size_t sofar;
+  ssize_t sofar;
 
   int yes = 1;
   int ret;
@@ -88,8 +88,8 @@ int main(int argc, char *argv[]) {
     return 1;
   }
 
-  printf("message size: %i octets\n", size);
-  printf("roundtrip count: %lli\n", count);
+  printf("message size: %li octets\n", size);
+  printf("roundtrip count: %li\n", count);
 
   if (!fork()) { /* child */
 
@@ -194,7 +194,7 @@ int main(int argc, char *argv[]) {
 
 #endif
 
-    printf("average latency: %lli ns\n", delta / (count * 2));
+    printf("average latency: %li ns\n", delta / (count * 2));
   }
 
   return 0;
